Only build otel on nodejs runtime
