+++ apps/dashboard/src/app/QueryProvider.tsx
@@ -3,7 +3,7 @@
 import { env } from "@/lib/env"
 import { TRPCProvider } from "@/lib/trpc"
 import type { AppRouter } from "@dotkomonline/gateway-trpc"
-import { getBrowserLogger } from "@dotkomonline/logger"
+import { getBrowserLogger } from "@dotkomonline/logger/browser"
 import { useSession } from "@dotkomonline/oauth2/react"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { type CreateTRPCClientOptions, createTRPCClient, httpBatchLink, loggerLink } from "@trpc/client"
