+++ apps/dashboard/src/instrumentation.ts
@@ -1,10 +1,12 @@
-import { getLogger, getResource, startOpenTelemetry } from "@dotkomonline/logger"
 import * as Sentry from "@sentry/nextjs"
 
-const logger = getLogger("monoweb-dashboard/instrumentation")
-
-export function register() {
-  const resource = getResource("monoweb-dashboard")
+export async function register() {
+  if (process.env.NEXT_RUNTIME !== "node") {
+    return
+  }
+  const { getLogger, getResource, startOpenTelemetry } = await import("@dotkomonline/logger")
+  const logger = getLogger("monoweb-web/instrumentation")
+  const resource = getResource("monoweb-web")
   startOpenTelemetry(resource)
 
   if (process.env.SENTRY_DSN !== undefined) {
