+++ apps/web/src/utils/trpc/QueryProvider.tsx
@@ -2,7 +2,7 @@
 
 import { env } from "@/env"
 import type { AppRouter } from "@dotkomonline/gateway-trpc"
-import { getBrowserLogger } from "@dotkomonline/logger"
+import { getBrowserLogger } from "@dotkomonline/logger/browser"
 import { useSession } from "@dotkomonline/oauth2/react"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { type CreateTRPCClientOptions, createTRPCClient, httpBatchLink, loggerLink } from "@trpc/client"
