+++ packages/logger/package.json
@@ -1,8 +1,16 @@
 {
   "name": "@dotkomonline/logger",
   "type": "module",
-  "main": "./src/index.ts",
-  "types": "./src/index.ts",
+  "exports": {
+    ".": {
+      "import": "./src/index.ts",
+      "types": "./src/index.ts"
+    },
+    "./browser": {
+      "import": "./src/browser.ts",
+      "types": "./src/browser.ts"
+    }
+  },
   "private": true,
   "scripts": {
     "clean": "rm -rf node_modules",
